annon_api v0.12.0 Annon.AcceptanceCase

This module defines the test case to be used by acceptance tests. It can allow run tests in async when each SQL.Sandbox connection will be binded to a specific test.

Functions

assert_status(response, status)

build_factory_params(factory, overrides \\ [])

build_invalid_plugin(plugin_name)

build_jwt_signature(signature)

build_jwt_token(payload, signature)

get_body(response)