annon_api v0.2.83 Annon.Configuration.Matcher

This module is responsible for configuration lookups for each request.

Functions

config_change(pid \\ __MODULE__)

Notifies cache adapter about configuration change.

init(args)

Initializes matcher state.

It accepts two options:

adapter - cache adapter.

cache_space - unique cache space key if multiple matcher processes are started.

match_request(pid \\ __MODULE__, scheme, method, host, port, path)

Returns API and associated Plugins by a request parameters.

This function receives adapter by casting GenServer and applies match in calling process context to reduce single-process bottleneck.

Adapter should expect their match functions to be read-concurrent from other processes.

start_link(opts, name \\ __MODULE__)