annon_api v0.2.85 Annon.DataCase

This module defines the setup for tests requiring access to the application’s data layer.

You may define functions here to be used as helpers in your tests.

Finally, if the test case interacts with the database, it cannot be async. For this reason, every test runs inside a transaction which is reset at the beginning of the test unless the test case is marked as async.

Functions

changeset_errors(changeset)

Helper for returning list of errors in a changeset.

Examples

Given a User schema that lists :name as a required field and validates :password to be safe, it would return:

iex> changeset_errors(%Changeset{})
[password: "is unsafe", name: "is blank"]

You could then write your assertion like:

assert {:password, "is unsafe"} in changeset_errors(%Changeset{})