annon_api v0.12.0 Annon.Factories.Configuration

This module lists Configuration-related factories, a mean suitable for tests that involve preparation of DB data.

Functions

acl_plugin_factory()

api_factory()

api_request_factory()

auth_plugin_with_jwt_factory()

auth_plugin_with_oauth_factory()

build(factory_name, attrs \\ %{})

build_list(number_of_factories, factory_name, attrs \\ %{})

build_pair(factory_name, attrs \\ %{})

cors_plugin_factory()

create()

create(any) :: no_return

create(, )

create(any, any) :: no_return

create_list(, , )

create_list(any, any, any) :: no_return

create_pair(, )

create_pair(any, any) :: no_return

factory(factory_name)

factory(any) :: no_return

Raises a helpful error if no factory is defined.

fields_for(factory_name, attrs \\ %{})

idempotency_plugin_factory()

insert(already_built_record)

insert(factory_name, attrs \\ %{})

insert_list(number_of_records, factory_name, attrs \\ %{})

insert_pair(factory_name, attrs \\ %{})

ip_restriction_plugin_factory()

params_for(factory_name, attrs \\ %{})

params_with_assocs(factory_name, attrs \\ %{})

proxy_plugin_factory()

string_params_for(factory_name, attrs \\ %{})

string_params_with_assocs(factory_name, attrs \\ %{})

ua_restriction_plugin_factory()

validator_plugin_factory()