annon_api v0.2.44 Annon.Helpers.Response

This is a helper module for dispatching requests.

It’s used by Annon.ManagementAPI.Render helpers and places where we want to return an error.

Functions

build_upstream_error(reason)

halt(conn)

Halt the connection.

Delegates to a Plug.Conn.halt/1 function.

send(conn, atom)

send(resource, conn, status)

Send request to a API consumer.

You may need to halt connection after calling it, if you want to stop rest of plugins from processing rests.

send_error(conn, arg2)

Send error by a EView.ErrorView template to a API consumer and halt connection.

send_error(conn, atom, root_object_name)

send_validation_error(conn, invalid)