annon_api v0.2.48 Annon.Plugins.APILoader

This plugin should be first in plugs pipeline, because it’s used to fetch all settings and decide which ones should be applied for current consumer request.

Functions

find_matching_method(apis, method)

find_matching_path(apis, path)

find_plugin_settings(plugins)

get_config(conn)

init(opts)