annon_api v0.2.77 Annon.Requests.LogWriter

This module provides interface to synchronously or asynchronously create Log Entries.

Functions

create_request(attrs)

Creates a Request.

Examples

iex> create_request(%{field: value})
{:ok, %Annon.Requests.Request{}}

iex> create_request(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_request_async(attrs, opts \\ [])

Asynchronously creates a Request. If request insert operations is failed, error is logged to a console.

Accepts optional opts as Keyword:

name - name or pid of a worker process.

Examples

iex> create_request(%{field: value})
:ok

iex> create_request(%{field: bad_value})
{:error, %Ecto.Changeset{}}

start_link(opts \\ [])

Start a LogWriter worker.

start_link accepts optional opts as Keyword:

name - name for a worker process;

subscriber - name or pid of a process that wants to be notified when log is written with create_request_async/1.