annon_api v0.12.0 Annon.Requests.Sandbox

A plug to allow concurrent, transactional acceptance tests with Annon.Requests module.

Functions

call(conn, sandbox)

init(opts \\ [])

metadata_for(repo_or_repos, pid)

metadata_for(Ecto.Repo.t | [Ecto.Repo.t], pid) :: map

Returns metadata to associate with the session to allow the endpoint to acces the database connection checked out by the test process.