gateway v0.2.3 Gateway.AcceptanceCase

This module defines the test case to be used by acceptance tests. It can allow run tests in async when each SQL.Sandbox connection will be binded to a specific test.

Functions

assert_status(response, status)

build_factory_params(factory, overrides \\ [])

build_jwt_signature(signature)

build_jwt_token(payload, signature)

get_body(response)