gateway v0.1.83 Gateway.DB.Logger.Repo

This repo is used to store request and responses.

We recommend database with low write latency, because it will have up to 2 writes for each API call that is going trough Annon.

Also, if Idempotency plug is enabled and X-Idempotency-Key: <key> header is sent by a consumer, you can expect an additional read request.

Functions

aggregate(queryable, aggregate, field, opts \\ [])

Callback implementation for Ecto.Repo.aggregate/4.

all(queryable, opts \\ [])

Callback implementation for Ecto.Repo.all/2.

config()

Callback implementation for Ecto.Repo.config/0.

delete(struct, opts \\ [])

Callback implementation for Ecto.Repo.delete/2.

delete!(struct, opts \\ [])

Callback implementation for Ecto.Repo.delete!/2.

delete_all(queryable, opts \\ [])

Callback implementation for Ecto.Repo.delete_all/2.

get(queryable, id, opts \\ [])

Callback implementation for Ecto.Repo.get/3.

get!(queryable, id, opts \\ [])

Callback implementation for Ecto.Repo.get!/3.

get_by(queryable, clauses, opts \\ [])

Callback implementation for Ecto.Repo.get_by/3.

get_by!(queryable, clauses, opts \\ [])

Callback implementation for Ecto.Repo.get_by!/3.

in_transaction?()

Callback implementation for Ecto.Repo.in_transaction?/0.

insert(struct, opts \\ [])

Callback implementation for Ecto.Repo.insert/2.

insert!(struct, opts \\ [])

Callback implementation for Ecto.Repo.insert!/2.

insert_all(schema_or_source, entries, opts \\ [])

Callback implementation for Ecto.Repo.insert_all/3.

insert_or_update(changeset, opts \\ [])

Callback implementation for Ecto.Repo.insert_or_update/2.

insert_or_update!(changeset, opts \\ [])

Callback implementation for Ecto.Repo.insert_or_update!/2.

load(schema_or_types, data)

Callback implementation for Ecto.Repo.load/2.

one(queryable, opts \\ [])

Callback implementation for Ecto.Repo.one/2.

one!(queryable, opts \\ [])

Callback implementation for Ecto.Repo.one!/2.

page(queryable, paging, opts \\ [])

page(queryable :: Ecto.Query.t, paging :: Ecto.Paging.t, opts :: Keyword.t) ::
  {[Ecto.Schema.t], Ecto.Paging.t} |
  no_return

Fetches all entries from the data store matching the given query with applied Pagging.

May raise Ecto.QueryError if query validation fails.

Options

See the “Shared options” section at the Ecto.Repo module documentation.

Example

# Fetch 50 post titles
query = from p in Post,
     select: p.title

MyRepo.all(query, %Ecto.Paging{limit: 50})

paginate(queryable, paging)

Convert queriable to queryable with applied pagination rules from paging.

preload(struct_or_structs, preloads, opts \\ [])

Callback implementation for Ecto.Repo.preload/3.

query(sql, params \\ [], opts \\ [])

A convenience function for SQL-based repositories that executes the given query.

See Ecto.Adapters.SQL.query/3 for more information.

query!(sql, params \\ [], opts \\ [])

A convenience function for SQL-based repositories that executes the given query.

See Ecto.Adapters.SQL.query/3 for more information.

rollback(value)

rollback(term) :: no_return

Callback implementation for Ecto.Repo.rollback/1.

start_link(opts \\ [])

Callback implementation for Ecto.Repo.start_link/1.

stop(pid, timeout \\ 5000)

Callback implementation for Ecto.Repo.stop/2.

stream(queryable, opts \\ [])

Callback implementation for Ecto.Repo.stream/2.

transaction(fun_or_multi, opts \\ [])

Callback implementation for Ecto.Repo.transaction/2.

update(struct, opts \\ [])

Callback implementation for Ecto.Repo.update/2.

update!(struct, opts \\ [])

Callback implementation for Ecto.Repo.update!/2.

update_all(queryable, updates, opts \\ [])

Callback implementation for Ecto.Repo.update_all/3.