gateway v0.2.3 Gateway.Helpers.Plugin behaviour

This helper provides abstract interface to define plugin.

Example:

use Gateway.Helpers.Plugin,
  plugin_name: "my_plugin"

It will add

init/1 settings required by Plug behavior, that will pass opts to init/2 methods.

find_plugin_settings/1 method that allows to find plugin settings and make sure that it’s enabled.

Callbacks

find_plugin_settings(list)

find_plugin_settings([Gateway.DB.Schemas.Plugin.t]) ::
  nil |
  Gateway.DB.Schemas.Plugin.t

Find plugin settings in plugins list. Returns nil if plugin is not found.