gateway v0.1.70 Gateway.Helpers.Render

Use this helpers when you want to render result in a controllers.

Functions

render_change(tuple, conn, status \\ 200)

This render should be used for Repo.create and Repo.update results.

render_collection(resources, conn)

This render should be used for Repo.all results.

render_delete(arg, conn)

This render should be used for Repo.delete_all results.

It will throw an error if you tried to delete more than one record in a DB.

render_schema(resource, conn)

This render should be used for Repo.one results.