gateway v0.1.53 Gateway.Helpers.Response

This is a helper module for dispatching requests.

It’s used by Gateway.Helpers.Render helpers and places where we want to return an error.

Functions

halt(conn)

Halt the connection.

Delegates to a Plug.Conn.halt/1 function.

send(resource, conn, status)

Send request to a API consumer.

You may need to halt connection after calling it, if you want to stop rest of plugins from processing rests.

send_error(conn, arg2)

Send error by a EView.ErrorView template to a API consumer and halt connection.