gateway v0.1.69 Gateway.Plugins.IPRestriction

IP Restriction plugin.

It allows to white/black-list consumers by a IP addresses. Also you can use it in Consumer plugin settings overrides to limit IP’s from which specific consumer can request your API.

Functions

find_plugin_settings(plugins)

init(opts)