gateway v0.1.64 Phoenix.Ecto.SQL.Sandbox

A plug to allow concurrent, transactional acceptance tests with Ecto.Adapters.SQL.Sandbox.

Example

This plug should only be used during tests. First, set a flag to enable it in config/test.exs:

config :your_app, sql_sandbox: true

And use the flag to conditionally add the plug to lib/your_app/endpoint.ex:

if Application.get_env(:your_app, :sql_sandbox) do
  plug Phoenix.Ecto.SQL.Sandbox
end

It’s important that this is at the top of endpoint.ex, before any other plugs.

Then, within an acceptance test, checkout a sandboxed connection as before. Use metadata_for/2 helper to get the session metadata to that will allow access to the test’s connection. Here’s an example using Hound:

use Hound.Helpers

setup do
  :ok = Ecto.Adapters.SQL.Sandbox.checkout(YourApp.Repo)
  metadata = Phoenix.Ecto.SQL.Sandbox.metadata_for(YourApp.Repo, self())
  Hound.start_session(metadata: metadata)
end

Functions

call(conn, sandbox)

init(opts \\ [])

metadata_for(repo_or_repos, pid)

metadata_for(Ecto.Repo.t | [Ecto.Repo.t], pid) :: map

Returns metadata to associate with the session to allow the endpoint to acces the database connection checked out by the test process.