gateway v0.1.71 gateway_tasks

Nice way to apply migrations inside a released application.

For a containerized version of Annon you can set APP_MIGRATE=true environment variable to run this tasks.

Example:

./bin/$APP_NAME command "gateway_tasks" migrate!

Functions

migrate!()