annon_api v0.12.0 Annon.Requests.LogWriter

This module provides interface to synchronously or asynchronously create Log Entries.

Functions

create_request(attrs)

Creates a Request.

Examples

iex> create_request(%{field: value})
{:ok, %Annon.Requests.Request{}}

iex> create_request(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_request_async(attrs)

Asynchronously creates a Request. If request insert operations is failed, error is logged to a console.

Examples

iex> create_request(%{field: value})
:ok

iex> create_request(%{field: bad_value})
{:error, %Ecto.Changeset{}}

start_link()

Start a LogWriter worker.

subscribe(name_or_pid)

unsubscribe(name_or_pid)